+++ src/core/Icon/styles.css
@@ -1,15 +1,21 @@
-.base {
-    composes: boxmodel from '../mixins/globals.css';
+.reset {
+    all: initial;
+    box-sizing: border-box;
+}
 
+.base {
+    display: inline-block;
     content: ' ';
     font-family: 'Material Icons';
     font-weight: normal;
     font-style: normal;
-    font-size: 1.5rem;
+    font-size: 24px;
+    padding: 0;
+    margin: 0;
+
     line-height: 1;
     letter-spacing: normal;
     text-transform: none;
-    display: inline-block;
     white-space: nowrap;
     word-wrap: normal;
     direction: ltr;
@@ -18,24 +24,11 @@
     -webkit-font-smoothing: antialiased;
 }
 
-.menu-icon {
-    pointer-events: none;
-    vertical-align: middle;
-    font-size: 16px;
-}
-
-.button-icon {
-    pointer-events: none;
-    vertical-align: middle;
-    margin-right: 10px;
-    font-size: 26px;
-}
-
 .align-icon .material-icons {
     pointer-events: none;
     line-height: 0.9375rem;
     height: 0.9375rem;
-    width: 1.5rem;
+    width: 24pxrem;
 }
 .align-icon .material-icons:first-child {
     margin-right: 0.75rem;
