+++ src/core/LinearProgress/styles.css
@@ -1,18 +1,23 @@
-/**
- * Adapted from https://materializecss.com/preloader.html
- */
+.reset {
+    all: initial;
+    box-sizing: border-box;
+}
 
-/* rtl:begin:ignore */
 .base {
-    composes: boxmodel font from '../mixins/globals.css';
-
     position: relative;
     display: block;
-    overflow: hidden;
+
     width: 100%;
     height: 4px;
     margin: 0;
     background-color: #eaf4ff;
+    overflow: hidden;
+    overflow-x: hidden;
+    overflow-y: hidden;
+}
+
+.progress {
+    background-color: #1976d2;
 }
 
 .progress.determinate {
@@ -20,14 +25,10 @@
     top: 0;
     left: 0;
     bottom: 0;
-    background-color: #1976d2;
     transition: width 0.3s linear;
 }
 
-.progress.indeterminate {
-    background-color: #1976d2;
-}
-
+/* rtl:ignore */
 .progress.indeterminate:before {
     position: absolute;
     top: 0;
@@ -40,6 +41,7 @@
     will-change: left, right;
 }
 
+/* rtl:ignore */
 .progress.indeterminate:after {
     position: absolute;
     top: 0;
@@ -53,19 +55,7 @@
     will-change: left, right;
 }
 
-[dir='rtl'] .base .progress.determinate {
-    left: auto;
-    right: 0;
-}
-
-[dir='rtl'] .base .progress.indeterminate:before {
-    animation-name: anim-indeterminate-rtl;
-}
-
-[dir='rtl'] .base .progress.indeterminate:after {
-    animation-name: anim-indeterminate-short-rtl;
-}
-
+/* rtl:ignore */
 @keyframes anim-indeterminate {
     0% {
         left: -35%;
@@ -83,6 +73,7 @@
     }
 }
 
+/* rtl:ignore */
 @keyframes anim-indeterminate-short {
     0% {
         left: -200%;
@@ -99,38 +90,3 @@
         right: -8%;
     }
 }
-
-@keyframes anim-indeterminate-rtl {
-    0% {
-        left: 100%;
-        right: -35%;
-    }
-
-    60% {
-        left: -90%;
-        right: 100%;
-    }
-
-    100% {
-        left: -90%;
-        right: 100%;
-    }
-}
-
-@keyframes anim-indeterminate-short-rtl {
-    0% {
-        left: 100%;
-        right: -200%;
-    }
-
-    60% {
-        left: -8%;
-        right: 107%;
-    }
-
-    100% {
-        left: -8%;
-        right: 107%;
-    }
-}
-/* rtl:end:ignore */
