+++ src/core/SplitButton/styles.js
@@ -1,0 +1,5 @@
+import classNames from '../../utils/css'
+import styles from './styles.json'
+import './styles.css'
+
+export default classNames(styles)
