+++ src/core/helpers/Divider/styles.css
@@ -1,4 +1,8 @@
 .base {
+    all: initial;
+    box-sizing: border-box;
+
+    display: inline-block;
     width: 100%;
     height: 1px;
     background-color: #e0e0e0;
